Rename HUD to follow PascalCase.
