+++ src/index.ts
@@ -9,7 +9,7 @@ import LocationSelector from "./components/LocationSelector.vue";
 import FundingAcknowledgement from "./components/FundingAcknowledgement.vue";
 import CreditLogos from "./components/CreditLogos.vue";
 import GeolocationButton from "./components/GeolocationButton.vue";
-import wwtHUD from "./components/wwtHUD.vue";
+import WwtHUD from "./components/WwtHUD.vue";
 import Gallery from "./components/Gallery.vue";
 
 
@@ -26,7 +26,7 @@ export {
   FundingAcknowledgement,
   CreditLogos,
   GeolocationButton,
-  wwtHUD,
+  WwtHUD,
   PositionCoords,
   useGeolocation,
   useWWTKeyboardControls,
